fixed: "StravaResponse [message=com.google.gson.JsonSyntaxException: java.lang.NumberFormatException: Expected an int but was 2153630452 at line 1 column 211 path $[0].id, errors=null"

+++ Strava API v3/src/javastrava/api/v3/model/StravaActivity.java
@@ -21,11 +21,11 @@ import com.google.gson.annotations.SerializedName;
  * @author Dan Shannon
  *
  */
-public class StravaActivity implements StravaCacheable<Integer>{
+public class StravaActivity implements StravaCacheable<Long>{
 	/**
 	 * Strava's unique identifier for the activity
 	 */
-	private Integer id;
+	private Long id;
 	/**
 	 * State of the resource (summary, detailed, etc.)
 	 */
@@ -787,7 +787,7 @@ public class StravaActivity implements StravaCacheable<Integer>{
 	 * @return the id
 	 */
 	@Override
-	public Integer getId() {
+	public Long getId() {
 		return this.id;
 	}
 	/**
@@ -1198,7 +1198,7 @@ public class StravaActivity implements StravaCacheable<Integer>{
 	/**
 	 * @param id the id to set
 	 */
-	public void setId(final Integer id) {
+	public void setId(final Long id) {
 		this.id = id;
 	}
 	/**
